+++ src/resources/containers/container.ts
@@ -63,6 +63,16 @@ export interface Container extends Resource<ContainerMetas> {
   events: Events<ContainerEvent>;
 }
 
+export type NewContainer = {
+  name: string;
+  environment_id: ResourceId;
+  image_id: ResourceId;
+  stateful: boolean;
+  annotations: Record<string, any> | null;
+  config: Config;
+  volumes?: VolumeSummary[];
+};
+
 export interface ContainerIncludes {
   creators?: CreatorIncludes;
   images?: {